詳細画面実装

+++ src/main/java/com/example/blog/repository/BlogRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.blog.model.Blog;
 
-public interface BlogRepository extends JpaRepository<Blog, Integer> {
+public interface BlogRepository extends JpaRepository<Blog, Long> {
 	
 }
